+++ util/wolfeAR_start.lua
@@ -1,11 +1,8 @@
 -- Feed waypoints to script based on "userclass.lua" by Wolf Labs (@wolfe_br)
 -- tobitege: customisations to only run AR waypointing (outside of ArchHud)
 
-local uclamp = utils.clamp
-
 local enableWaypointer = true --export: Enable waypoint AR display. Default: checked.
---local wpRenderLimit = 3 --export: The number of kilometers above which distance waypoints are not rendered; 0 means display all
---wpRenderLimit = uclamp(wpRenderLimit, 0, 1000000) -- max 500 SU
+local wpRenderLimit = 5 --export: The number of kilometers above which distance waypoints are not rendered; 0 means display all
 
 if not enableWaypointer then
     WP_WOLF_ENABLED = false
@@ -13,6 +10,7 @@ if not enableWaypointer then
 end
 
 if WM and WolfAR then
+    WolfAR.setRenderLimit(wpRenderLimit)
     for _,p in ipairs(WM:getSorted()) do
         local s = PM.MapPos2String(p.mapPos)
         local wPos = PM.MapPosToWorldVec3(s);